Created Parameter class in core-common, to be used by both client and server
