+++ core-common/src/main/java/org/glassfish/jersey/model/internal/spi/ParameterServiceProvider.java
@@ -1,0 +1,40 @@
+/*
+ * Copyright (c) 2019 Oracle and/or its affiliates. All rights reserved.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License v. 2.0, which is available at
+ * http://www.eclipse.org/legal/epl-2.0.
+ *
+ * This Source Code may also be made available under the following Secondary
+ * Licenses when the conditions for such availability set forth in the
+ * Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
+ * version 2 with the GNU Classpath Exception, which is available at
+ * https://www.gnu.org/software/classpath/license.html.
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
+ */
+
+package org.glassfish.jersey.model.internal.spi;
+
+import org.glassfish.jersey.model.Parameter;
+
+import java.util.Map;
+
+/**
+ * Implementation of this service is used by any Parameter implementation
+ * to overcome the inheritance of static methods.
+ */
+public interface ParameterServiceProvider {
+    /**
+     * A map of supported Parameter Annotations. The maps from each found {@link ParameterServiceProvider}
+     * are combined to provide a list of Parameter Annotations for the given {@link Parameter}.
+     * @return A map of subset of Parameter Annotations known to the ParameterServiceProvider
+     */
+    Map<Class, Parameter.ParamAnnotationHelper> getParameterAnnotationHelperMap();
+
+    /**
+     * Get a factory to be used to instantiate the {@link Parameter}
+     * @return {@code Parameter} instantiating factory
+     */
+    Parameter.ParamCreationFactory<? extends Parameter> getParameterCreationFactory();
+}
