+++ core-common/src/main/resources/META-INF/services/org.glassfish.jersey.model.internal.spi.ParameterServiceProvider
@@ -1,0 +1,1 @@
+org.glassfish.jersey.model.Parameter$ParameterService
